Break dependendcy on transducers-js
The transducers-js dependency was blowing up our module size, and we
implemented nearly all of transducers-js ourselves. All we needed to add
was the actual `transduce` implementation and the remaining transforms.
In particular, this means we no longer need to rely on our index
argument hack that assumes `transducer-js`'s transforms visit each
element a single time in order.