Refactor to enable tree shaking, change API a bit
Big restructuring of project to create a tree-shakeable API, in addition
to the existing chain API. This means creating standalone transducers, a
compose function, and a transduce function, similar to those seen by
other libraries. Furthermore, switch to ES2015 modules, mark
`"sideEffects": false` in `package.json`, and add a script to replace
`/** @class */` with `/**@__PURE__*/` in the generated code (see
`markClassesPure.sh` for details).

Additionally, remove the ability treat arbitrary objects as instances of
`Iterable<[string, T]>` and rename `stringJoin()` to `joinToString()`.