Implement groupBy, remove index argument
Implement the wildly exciting `toObjectGroupBy` and `toMapGroupBy`
reductions, which allow you to pass a transformer to say how the
reduction should occur, e.g. grouping into arrays or into counts.

For this to work, all transformers must be "pure" in that they do not
hold state of their own (but the `TResult` type which is handled by its
functions is allowed to be mutable). Unfortunately, the indexing on all
functions causes performance issues when put into this framework, and
even without it caused a smaller performance hit. Take out the indexing,
but we will need to add a `mapIndexed` function to replace it.